+++ tests/com/twitter/ConformanceTest.java
@@ -10,7 +10,17 @@ import java.io.FileNotFoundException;
 import junit.framework.TestCase;
 import org.ho.yaml.Yaml;
 
-public class ConformanceTest extends ConformanceBase {
+public class ConformanceTest extends TestCase {
+
+  private static final String CONFORMANCE_DIR_PROPERTY = "conformance.dir";
+  protected static final String KEY_DESCRIPTION = "description";
+  protected static final String KEY_INPUT = "text";
+  protected static final String KEY_EXPECTED_OUTPUT = "expected";
+  protected static final String KEY_HIGHLIGHT_HITS = "hits";
+  protected File conformanceDir;
+  protected Extractor extractor = new Extractor();
+  protected Autolink linker = new Autolink();
+  protected HitHighlighter hitHighlighter = new HitHighlighter();
 
   public void testMentionsExtractor() throws Exception {
     File yamlFile = new File(conformanceDir, "extract.yml");
@@ -128,4 +138,28 @@ public class ConformanceTest extends ConformanceBase {
     }
   }
 
+  public void setUp() {
+    assertNotNull("Missing required system property: " + CONFORMANCE_DIR_PROPERTY, System.getProperty(
+            CONFORMANCE_DIR_PROPERTY));
+    conformanceDir = new File(System.getProperty(CONFORMANCE_DIR_PROPERTY));
+    assertTrue("Conformance directory " + conformanceDir + " is not a directory.", conformanceDir.isDirectory());
+
+    assertNotNull("No extractor configured", extractor);
+    assertNotNull("No autolinker configured", linker);
+    linker.setNoFollow(false);
+  }
+
+  protected void autolink(List<Map> testCases) {
+    for (Map testCase : testCases) {
+      assertEquals((String)testCase.get(KEY_DESCRIPTION),
+                   (String)testCase.get(KEY_EXPECTED_OUTPUT),
+                   linker.autoLink((String) testCase.get(KEY_INPUT)));
+    }
+  }
+
+  protected List loadConformanceData(File yamlFile, String testType) throws FileNotFoundException {
+    Map fullConfig = (Map) Yaml.load(yamlFile);
+    Map testConfig = (Map)fullConfig.get("tests");
+    return (List)testConfig.get(testType);
+  }
 }
